Add SourcifyFilePathBackfiller
